// src/github/types.ts
export type Event =
  | 'branch_protection_rule'
  | 'check_run'
  | 'check_suite'
  | 'create'
  | 'delete'
  | 'deployment'
  | 'deployment_status'
  | 'discussion'
  | 'discussion_comment'
  | 'fork'
  | 'gollum'
  | 'issue_comment'
  | 'issues'
  | 'label'
  | 'merge_group'
  | 'milestone'
  | 'page_build'
  | 'project'
  | 'project_card'
  | 'project_column'
  | 'public'
  | 'pull_request'
  | 'pull_request_review'
  | 'pull_request_review_comment'
  | 'pull_request_target'
  | 'push'
  | 'registry_package'
  | 'release'
  | 'status'
  | 'watch'
  | 'workflow_call'
  | 'workflow_dispatch'
  | 'workflow_run'
  | 'repository_dispatch';
// ...
export type StringContainingExpressionSyntax = string;
// ...
export interface HttpsJsonSchemastoreOrgGithubWorkflowJson {
// ...
  env?:
    | {
        [k: string]: string | number | boolean;
      }
    | StringContainingExpressionSyntax;
// ...
}